# api/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cat Activities Monitor API",
    "description": "\n    A comprehensive FastAPI-based system for monitoring cats using YOLO object detection and activity analysis.    ",
    "license": {
      "name": "MIT License",
      "url": "https://github.com/biancarosa/cat-activities-monitor/blob/main/LICENSE"
    },
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://your-domain.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/system/status": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Get System Status",
        "description": "Get comprehensive system status including configuration, services, and runtime information.",
        "operationId": "get_system_status_system_status_get",
        "responses": {
          "200": {
            "description": "System status retrieved successfully",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "status": "running",
                  "version": "2.0.0",
                  "uptime": "2h 15m 30s",
                  "configuration_loaded": true,
                  "background_task_running": true,
                  "database_connected": true,
                  "ml_model_loaded": true,
                  "enabled_cameras": 3,
                  "total_cameras": 5
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/system/config": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Get Current Configuration",
        "description": "Retrieve the current system configuration including camera sources, YOLO settings, and detection parameters.",
        "operationId": "get_current_config_system_config_get",
        "responses": {
          "200": {
            "description": "Configuration retrieved successfully",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "global_": {
                    "ml_model_config": {
                      "model": "yolo11n.pt",
                      "confidence_threshold": 0.5,
                      "target_classes": [
                        15,
                        16
                      ]
                    },
                    "default_interval_seconds": 60,
                    "max_concurrent_fetches": 3,
                    "timeout_seconds": 30
                  },
                  "images": [
                    {
                      "name": "living_room",
                      "url": "http://camera1.local/snapshot.jpg",
                      "enabled": true,
                      "interval_seconds": 30
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "No configuration loaded"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/system/config/reload": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Reload Configuration",
        "description": "Reload the system configuration from the config file.",
        "operationId": "reload_configuration_system_config_reload_post",
        "responses": {
          "200": {
            "description": "Configuration reloaded successfully",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "message": "Configuration reloaded successfully",
                  "total_images": 5,
                  "enabled_images": 3,
                  "ml_model": "yolo11n.pt"
                }
              }
            }
          },
          "500": {
            "description": "Failed to reload configuration"
          }
        }
      }
    },
    "/system/logs": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Get Recent Logs",
        "description": "Retrieve recent log entries from the application log file.",
        "operationId": "get_recent_logs_system_logs_get",
        "parameters": [
          {
            "name": "lines",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Lines"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Log entries retrieved successfully",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "logs": [
                    "2024-01-15 10:30:00 - INFO - Background image fetcher started",
                    "2024-01-15 10:29:45 - INFO - ML model loaded: yolo11n.pt"
                  ],
                  "total_lines": 2
                }
              }
            }
          },
          "500": {
            "description": "Failed to read log file"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}